refactor(sandbox): reorganize Windows sandbox files to match Linux and macOS structure
