+++ packages/core/src/sandbox/windows/GeminiSandbox.cs
